Add ability to poll RocksDB for a pending key
Since there is a race condition between the oracle and other nodes (see
the note about it in the README), add configuration options to poll
RocksDB on a GET request - number of retries and sleep period between
rertries.

A more optimal solution could be introduced in a future release. Client
retries could also be added.

A test for the race condition will be added in a subsequent commit.